requirements.txt
aiodns==2.0.0
aiogram==2.4
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
Babel==2.7.0
brotlipy==0.7.0
cchardet==2.1.5
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
dnspython==1.16.0
idna==2.8
multidict==4.7.1
pycares==3.1.0
pycparser==2.19
pymongo==3.10.0
python-dotenv==0.10.3
pytz==2019.3
ujson==1.35
yarl==1.4.2
